Selection - implement `SelectCommand`
This command is designed to replace the functionality of the SelectorNode. The primary difference is that this command expects that the animation data is provided as one of it's argument. Providing the animation data through the argument is more flexible and less complex than the previous design of extracting the animation data.